+++ src/main/kotlin/com/github/rafaelldi/tyeplugin/cli/TyeCliClient.kt
@@ -32,6 +32,7 @@ class TyeCliClient {
         val verbosity: String,
         val tags: String?,
         val debug: String?,
+        val framework: String?,
         val logsProvider: String,
         val logsProviderUrl: String?,
         val tracesProvider: String,
@@ -75,6 +76,7 @@ class TyeCliClient {
         if (options.verbosity != DEFAULT_VERBOSITY) cliBuilder.setVerbosity(options.verbosity)
         if (!options.tags.isNullOrBlank()) cliBuilder.setTags(options.tags)
         if (!options.debug.isNullOrBlank()) cliBuilder.setDebug(options.debug)
+        if (!options.framework.isNullOrBlank()) cliBuilder.setFramework(options.framework)
         if (options.logsProvider != DEFAULT_LOGS_PROVIDER) cliBuilder.setLogs(
             options.logsProvider,
             options.logsProviderUrl