+++ src/main/kotlin/com/github/rafaelldi/tyeplugin/remoteServer/deployment/TyeDeploymentConfiguration.kt
@@ -16,6 +16,7 @@ class TyeDeploymentConfiguration : DeploymentConfigurationBase<TyeDeploymentConf
     var verbosityArgument = Verbosity.INFO
     var tagsArgument: String? = null
     var debugArgument: String? = null
+    var frameworkArgument: String? = null
     var logsProvider = LogsProvider.NONE
     var logsProviderUrl: String? = null
     var tracesProvider = TracesProvider.NONE